fix(): prevent listener firing because of object patching
Reordering the patching of individual elements before
the patching of prototypes allows functions on the
prototypes to be called during that process without
triggering listeners. Also, tests on the patching
of these individual elements must add test elements
to the DOM or else they will not be patched during
the retrieval.

Closes #2