docs(domInterceptor): update comments for proxy methods to explain they are not currently used
At this point, the use of proxies is considered experimental javascript. Therefore the methods
used to create proxies of DOM elements that will call the listener function when their elements
are manipulated are left as useful methods for the purpose of detecting DOM manipulation, but
they are not used by addListener or removeListener. Instead, the overall implementation still
works with the current browser technology.