fix(domInterceptor): update lineNumber defaults to fix error printing
The HintLog module prints messages with lineNumbers that are found
by creating new errors and spliting the stacktrace. This splitting
helps to find the error that is in the users code rather than
printing a stacktrace of the inner workings of an ngHint module.
This means HintLog must know the correct line to print for each
module. This fix sets the default line for domInterceptor to six.