fix(safari): add conditional testing and implementation to support Safari
Safari does not support patching individual HTML elements or the EventTarget property. The current
listener strategy does not use patching individual HTML elements because this was deemed
unnecessary due to performance and issues of robustness. Hence no implementation change is
necessary to alter the HTML patching for Safari. Only the tests are altered to be conditional and
comments are clarified to note Safari's behavior. Conditional logic is added to both tests and
implementation regarding EventTarget.